requirements.txt
streamlit>=1.22.0
python-dotenv>=1.0.0
apscheduler>=3.10.1
plyer>=2.1.0
google-auth>=2.18.1
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.86.0
firebase-admin>=6.1.0
SpeechRecognition>=3.10.0
pyaudio>=0.2.13
pandas>=2.0.2
numpy>=1.24.3
pillow>=9.5.0
PyJWT>=2.7.0
requests>=2.31.0
beautifulsoup4>=4.12.2
python-dateutil>=2.8.2
